fix: harden tls remediation detection
